fix: add-notebook flag

+++ scripts/chromedriver.sh
@@ -1,35 +1,47 @@
 #!/bin/sh
 
+notebook=false
 version=92.0.4515.107
 addr=64
 
-while getopts v:a: flag
+while getopts v:a:n: flag
 do
     # shellcheck disable=SC2220
     case "${flag}" in
         v) version=${OPTARG};;
         a) addr=${OPTARG};;
+        n) notebook=true;;
     esac
 done
 
-sudo apt-get update
-sudo apt-get install -y unzip xvfb libxi6 libgconf-2-4
+if "$notebook"; then
 
-sudo apt-get install default-jdk
+  sudo apt-get update
+  sudo apt-get install -y unzip xvfb libxi6 libgconf-2-4
 
-sudo curl -sS -o - https://dl-ssl.google.com/linux/linux_signing_key.pub | apt-key add
-sudo echo "deb [arch=amd64]  http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google-chrome.list
-sudo apt-get -y update
-sudo apt-get -y install google-chrome-stable
+  sudo apt-get install default-jdk
 
-# https://chromedriver.storage.googleapis.com/index.html
+  sudo curl -sS -o - https://dl-ssl.google.com/linux/linux_signing_key.pub | apt-key add
+  sudo echo "deb [arch=amd64]  http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google-chrome.list
+  sudo apt-get -y update
+  sudo apt-get -y install google-chrome-stable
 
-wget https://chromedriver.storage.googleapis.com/"$version"/chromedriver_linux"$addr".zip
-# shellcheck disable=SC2086
-unzip chromedriver_linux$addr.zip
+  # https://chromedriver.storage.googleapis.com/index.html
 
-sudo mv chromedriver /usr/bin/chromedriver
-sudo chown root:root /usr/bin/chromedriver
-sudo chmod +x /usr/bin/chromedriver
+  wget https://chromedriver.storage.googleapis.com/"$version"/chromedriver_linux"$addr".zip
+  # shellcheck disable=SC2086
+  unzip chromedriver_linux$addr.zip
 
-sudo rm chromedriver_linux"$addr".zip
+  sudo mv chromedriver /usr/bin/chromedriver
+  sudo chown root:root /usr/bin/chromedriver
+  sudo chmod +x /usr/bin/chromedriver
+
+  sudo rm chromedriver_linux"$addr".zip
+
+else
+
+  apt-get update
+  apt install chromium-chromedriver
+  cp /usr/lib/chromium-browser/chromedriver /usr/bin
+
+fi
